Add files via upload
SQL File Description:
This SQL file contains the script to create tables specifically designed for data visualization in Tableau. The tables are structured to optimize data analysis and visualization in Tableau dashboards.

Purpose:
The purpose of this script is to provide a standardized and efficient way to create the necessary tables for visualizing COVID-19 data in Tableau. These tables are designed to support various data visualizations, including charts, graphs, and interactive dashboards.